delete old sql code
