--- integrations/sql/src/main/kotlin/com/xebia/functional/xef/sql/jdbc/JdbcConfig_deprecated.kt
@@ -1,15 +1,0 @@
-package com.xebia.functional.xef.sql.jdbc
-
-import com.xebia.functional.xef.llm.Chat
-
-class JdbcConfig_deprecated(
-  val vendor: String,
-  val host: String,
-  val username: String,
-  val password: String,
-  val port: Int,
-  val database: String,
-  val model: Chat
-) {
-  fun toJDBCUrl(): String = "jdbc:$vendor://$host:$port/$database"
-}
